# Improve performance of constant aggregate window expression