@alamb 2025-06-03
🤖 `./gh_compare_branch_bench.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_branch_bench.sh) Running
Linux aal-dev 6.11.0-1013-gcp #13~24.04.1-Ubuntu SMP Wed Apr  2 16:34:16 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing constant_agg_window (cf3d1b05e50590c75cda563610d1e132a32d0acc) to 7248259a69b5c59f39d25f944dc1b54686d0eaf4 [diff](https://github.com/apache/datafusion/compare/7248259a69b5c59f39d25f944dc1b54686d0eaf4..cf3d1b05e50590c75cda563610d1e132a32d0acc)
BENCH_NAME=h2o_small_window
BENCH_COMMAND=cargo bench --bench h2o_small_window
BENCH_FILTER=
BENCH_BRANCH_NAME=constant_agg_window
Results will be posted here when complete