@andygrove 2025-06-06
I still see test failures:

```
2025-06-06T18:02:24.9752586Z - aggregate window function for all types *** FAILED *** (260 milliseconds)
2025-06-06T18:02:24.9755112Z   64 did not equal 1027 [64] did not equal [1027] (CometTestBase.scala:130)
```

and

```
2025-06-06T18:02:25.3200520Z - Windows support *** FAILED *** (342 milliseconds)
2025-06-06T18:02:25.3201792Z   Results do not match for query:

2025-06-06T18:02:25.3254822Z   == Results ==
2025-06-06T18:02:25.3255257Z   !== Correct Answer - 10 ==                                                                  == Spark Answer - 10 ==
2025-06-06T18:02:25.3256419Z    struct<count(_1) OVER (ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING):bigint>   struct<count(_1) OVER (ROWS BETWEEN UNBOUNDED PRECEDING AND UNBOUNDED FOLLOWING):bigint>
2025-06-06T18:02:25.3257476Z   ![10]                                                                                       [12]
2025-06-06T18:02:25.3257914Z   ![10]                                                                                       [12]
2025-06-06T18:02:25.3258358Z   ![10]                                                                                       [20]
2025-06-06T18:02:25.3258817Z   ![10]                                                                                       [20]
2025-06-06T18:02:25.3259269Z   ![10]                                                                                       [2]
2025-06-06T18:02:25.3259712Z   ![10]                                                                                       [2]
2025-06-06T18:02:25.3260149Z   ![10]                                                                                       [30]
2025-06-06T18:02:25.3260763Z   ![10]                                                                                       [30]
2025-06-06T18:02:25.3261200Z   ![10]                                                                                       [6]
2025-06-06T18:02:25.3261669Z   ![10]                                                                                       [6]
```

I suggest we revert this PR for now and then add more tests based on the failing tests in Spark/Comet so that we can have more confidence when the PR is updated.