@alamb 2025-06-06
> I suggest we revert this PR for now and then add more tests based on the failing tests in Spark/Comet so that we can have more confidence when the PR is updated.

Update:@andygrove  reverted the change on the branch-48 line: 
- https://github.com/apache/datafusion/pull/16307
However it is not reverted on main

Let's continue the discussion and debugging on
- https://github.com/apache/datafusion/issues/16308